Updated logic that was handling the protocol
Updated logic that was handling the protocol for custom domains and instead added an option to the add domain screen to select the protocol. Updated the code to use the protocol to build the link instead of relying on server vars since some may still want to use plain http. This also fixed the copy problem.